Fix toJSON() method
GET /get would sometimes return a string containing two JSON
objects, probably due to memory corruption. This probably happened
due to the use of a static variable in that method.

Now, callers need to pass their String as an input argument - this also
takes care of any memory leaks related to allocating on the heap in toJSON().

Also fix /trigger_restart: need to delay() a bit before restarting or
the restart won't happen.