Remove autotune feature
I never really used it, because it's not that useful for the coffee machine
- in my opinion.

Based on my understanding of the typical autotune features:

The autotune algorithms expect an initial steady state with a steady disturbance,
e.g. a constant amount of heat loss. In that scenario, they will tune the algorithm
to keep a steady temperature when the machine is idle.

That scenario is not really interesting: we are also interested in having
a quick recovery inter-shot and stable intra-shot temperatures.

That scenario is not really covered by the autotune algorithm. One can try
to introduce this disturbance during the autotune run, but I do not know
if that will be successful and repeatable enough. Additionally, the autotune
algorithm changes the setpoint itself (by a pre-configured amount) to
observe how the system behaves. By introducing external changes at
unpredictable times (i.e. remove the steady state by increasing the
loss of heat), one will potentially confuse the algorithm.

TL;DR: autotune is probably best used when heat loss is reasonably constant.